\name{green.data}
\alias{green.data}
\docType{data}
\title{
Symbiodinium sp. ITS2 OTUs from Orbicella franksi and O. faveolata 
}
\description{
Symbiodinium sp. ITS2 OTUs from Orbicella franksi and O. faveolata corals at the Flower Garden Banks. Two banks (east and west), two coral species, five OTUs.
}
\usage{data(green.data)}
\format{
  A data frame with 58 observations on the following 156 variables.
  \describe{
    \item{\code{species}}{a factor with levels \code{faveolata} \code{franksi}}
    \item{\code{bank}}{a factor with levels \code{east} \code{west}}
    \item{\code{sample}}{a factor with levels \code{EFAV117} \code{EFAV123} \code{EFAV128} \code{EFAV137} \code{EFAV140} \code{EFAV142} \code{EFAV143} \code{EFAV145} \code{EFAV150} \code{EFAV153} \code{EFAV154} \code{EFAV157} \code{EFAV161} \code{EFAV162} \code{EFRA101} \code{EFRA102} \code{EFRA105} \code{EFRA216} \code{EFRA218} \code{EFRA221} \code{EFRA222} \code{EFRA224} \code{EFRA231} \code{EFRA232} \code{EFRA233} \code{EFRA236} \code{EFRA82} \code{EFRA90} \code{EFRA91} \code{WFAV102} \code{WFAV103} \code{WFAV107} \code{WFAV111} \code{WFAV112} \code{WFAV188} \code{WFAV191} \code{WFAV89} \code{WFAV91} \code{WFAV94} \code{WFAV95} \code{WFAV97} \code{WFAV98} \code{WFAV99} \code{WFRA106} \code{WFRA107} \code{WFRA112} \code{WFRA114} \code{WFRA115} \code{WFRA116} \code{WFRA117} \code{WFRA198} \code{WFRA199} \code{WFRA200} \code{WFRA201} \code{WFRA202} \code{WFRA203} \code{WFRA205} \code{WFRA210}}
    \item{\code{H35JRAZ01A0C9J}}{a numeric vector}
    \item{\code{H35JRAZ01A0D1B}}{a numeric vector}
    \item{\code{H35JRAZ01A0RQO}}{a numeric vector}
    \item{\code{H35JRAZ01A1FKL}}{a numeric vector}
    \item{\code{H35JRAZ01A1GT3}}{a numeric vector}
    \item{\code{H35JRAZ01A1NMH}}{a numeric vector}
    \item{\code{H35JRAZ01A1SO3}}{a numeric vector}
    \item{\code{H35JRAZ01A1VUD}}{a numeric vector}
    \item{\code{H35JRAZ01A22NW}}{a numeric vector}
    \item{\code{H35JRAZ01A2ZJK}}{a numeric vector}
    \item{\code{H35JRAZ01A2ZYY}}{a numeric vector}
    \item{\code{H35JRAZ01A30IE}}{a numeric vector}
    \item{\code{H35JRAZ01A32BS}}{a numeric vector}
    \item{\code{H35JRAZ01A3DXY}}{a numeric vector}
    \item{\code{H35JRAZ01A3UJS}}{a numeric vector}
    \item{\code{H35JRAZ01A4BHG}}{a numeric vector}
    \item{\code{H35JRAZ01A4F3O}}{a numeric vector}
    \item{\code{H35JRAZ01A4G6O}}{a numeric vector}
    \item{\code{H35JRAZ01A4O8A}}{a numeric vector}
    \item{\code{H35JRAZ01A4U5O}}{a numeric vector}
    \item{\code{H35JRAZ01A5LTB}}{a numeric vector}
    \item{\code{H35JRAZ01A5O9U}}{a numeric vector}
    \item{\code{H35JRAZ01A6ITE}}{a numeric vector}
    \item{\code{H35JRAZ01A771F}}{a numeric vector}
    \item{\code{H35JRAZ01A7EP4}}{a numeric vector}
    \item{\code{H35JRAZ01A7SMY}}{a numeric vector}
    \item{\code{H35JRAZ01A8A59}}{a numeric vector}
    \item{\code{H35JRAZ01A8X2Y}}{a numeric vector}
    \item{\code{H35JRAZ01A93HV}}{a numeric vector}
    \item{\code{H35JRAZ01A9TZY}}{a numeric vector}
    \item{\code{H35JRAZ01AMZP0}}{a numeric vector}
    \item{\code{H35JRAZ01AN7JF}}{a numeric vector}
    \item{\code{H35JRAZ01ANHED}}{a numeric vector}
    \item{\code{H35JRAZ01ANUPT}}{a numeric vector}
    \item{\code{H35JRAZ01AO7GS}}{a numeric vector}
    \item{\code{H35JRAZ01AON2S}}{a numeric vector}
    \item{\code{H35JRAZ01APP5A}}{a numeric vector}
    \item{\code{H35JRAZ01APPG5}}{a numeric vector}
    \item{\code{H35JRAZ01APWEH}}{a numeric vector}
    \item{\code{H35JRAZ01APYLM}}{a numeric vector}
    \item{\code{H35JRAZ01AQIU4}}{a numeric vector}
    \item{\code{H35JRAZ01AR1R8}}{a numeric vector}
    \item{\code{H35JRAZ01AR8Q4}}{a numeric vector}
    \item{\code{H35JRAZ01ARD6M}}{a numeric vector}
    \item{\code{H35JRAZ01ARKYW}}{a numeric vector}
    \item{\code{H35JRAZ01ARSTD}}{a numeric vector}
    \item{\code{H35JRAZ01ARZ7L}}{a numeric vector}
    \item{\code{H35JRAZ01AS8ZF}}{a numeric vector}
    \item{\code{H35JRAZ01ASV28}}{a numeric vector}
    \item{\code{H35JRAZ01ATSK2}}{a numeric vector}
    \item{\code{H35JRAZ01ATZCY}}{a numeric vector}
    \item{\code{H35JRAZ01AU38T}}{a numeric vector}
    \item{\code{H35JRAZ01AUMOE}}{a numeric vector}
    \item{\code{H35JRAZ01AUTH5}}{a numeric vector}
    \item{\code{H35JRAZ01AUX8X}}{a numeric vector}
    \item{\code{H35JRAZ01AVB0Q}}{a numeric vector}
    \item{\code{H35JRAZ01AVFAX}}{a numeric vector}
    \item{\code{H35JRAZ01AVX70}}{a numeric vector}
    \item{\code{H35JRAZ01AWGQC}}{a numeric vector}
    \item{\code{H35JRAZ01AXA3Z}}{a numeric vector}
    \item{\code{H35JRAZ01AXGLD}}{a numeric vector}
    \item{\code{H35JRAZ01AXVIR}}{a numeric vector}
    \item{\code{H35JRAZ01AY2DE}}{a numeric vector}
    \item{\code{H35JRAZ01AY6KB}}{a numeric vector}
    \item{\code{H35JRAZ01AZFWS}}{a numeric vector}
    \item{\code{H35JRAZ01AZTST}}{a numeric vector}
    \item{\code{H35JRAZ01BB9FU}}{a numeric vector}
    \item{\code{H35JRAZ01BBITP}}{a numeric vector}
    \item{\code{H35JRAZ01BBSV6}}{a numeric vector}
    \item{\code{H35JRAZ01BD2HV}}{a numeric vector}
    \item{\code{H35JRAZ01BDBD6}}{a numeric vector}
    \item{\code{H35JRAZ01BDTVF}}{a numeric vector}
    \item{\code{H35JRAZ01BE95X}}{a numeric vector}
    \item{\code{H35JRAZ01BEHEK}}{a numeric vector}
    \item{\code{H35JRAZ01BEMYI}}{a numeric vector}
    \item{\code{H35JRAZ01BEP0J}}{a numeric vector}
    \item{\code{H35JRAZ01BFNRT}}{a numeric vector}
    \item{\code{H35JRAZ01BFVYJ}}{a numeric vector}
    \item{\code{H35JRAZ01BG87N}}{a numeric vector}
    \item{\code{H35JRAZ01BGG1A}}{a numeric vector}
    \item{\code{H35JRAZ01BGN2J}}{a numeric vector}
    \item{\code{H35JRAZ01BGUP9}}{a numeric vector}
    \item{\code{H35JRAZ01BH7VV}}{a numeric vector}
    \item{\code{H35JRAZ01BIHI3}}{a numeric vector}
    \item{\code{H35JRAZ03C08W3}}{a numeric vector}
    \item{\code{H35JRAZ03C0K1K}}{a numeric vector}
    \item{\code{H35JRAZ03C141R}}{a numeric vector}
    \item{\code{H35JRAZ03C16B4}}{a numeric vector}
    \item{\code{H35JRAZ03C1INY}}{a numeric vector}
    \item{\code{H35JRAZ03C1WD2}}{a numeric vector}
    \item{\code{H35JRAZ03C2344}}{a numeric vector}
    \item{\code{H35JRAZ03C2919}}{a numeric vector}
    \item{\code{H35JRAZ03C2CBI}}{a numeric vector}
    \item{\code{H35JRAZ03C2DMJ}}{a numeric vector}
    \item{\code{H35JRAZ03C2MH3}}{a numeric vector}
    \item{\code{H35JRAZ03C2PQY}}{a numeric vector}
    \item{\code{H35JRAZ03C32WG}}{a numeric vector}
    \item{\code{H35JRAZ03C3ANP}}{a numeric vector}
    \item{\code{H35JRAZ03C5910}}{a numeric vector}
    \item{\code{H35JRAZ03C5CKR}}{a numeric vector}
    \item{\code{H35JRAZ03C5G30}}{a numeric vector}
    \item{\code{H35JRAZ03C60I8}}{a numeric vector}
    \item{\code{H35JRAZ03C6AF1}}{a numeric vector}
    \item{\code{H35JRAZ03C6E0W}}{a numeric vector}
    \item{\code{H35JRAZ03C6LU3}}{a numeric vector}
    \item{\code{H35JRAZ03C6Q3V}}{a numeric vector}
    \item{\code{H35JRAZ03C75OM}}{a numeric vector}
    \item{\code{H35JRAZ03C7GCE}}{a numeric vector}
    \item{\code{H35JRAZ03C7P6J}}{a numeric vector}
    \item{\code{H35JRAZ03C8JJT}}{a numeric vector}
    \item{\code{H35JRAZ03C9MVI}}{a numeric vector}
    \item{\code{H35JRAZ03CX46J}}{a numeric vector}
    \item{\code{H35JRAZ03CX5BC}}{a numeric vector}
    \item{\code{H35JRAZ03CXVYS}}{a numeric vector}
    \item{\code{H35JRAZ03DA80G}}{a numeric vector}
    \item{\code{H35JRAZ03DALPN}}{a numeric vector}
    \item{\code{H35JRAZ03DAW5J}}{a numeric vector}
    \item{\code{H35JRAZ03DAXEI}}{a numeric vector}
    \item{\code{H35JRAZ03DC5UZ}}{a numeric vector}
    \item{\code{H35JRAZ03DCCVA}}{a numeric vector}
    \item{\code{H35JRAZ03DCGCE}}{a numeric vector}
    \item{\code{H35JRAZ03DCXOX}}{a numeric vector}
    \item{\code{H35JRAZ03DE7WW}}{a numeric vector}
    \item{\code{H35JRAZ03DED1S}}{a numeric vector}
    \item{\code{H35JRAZ03DEJ2S}}{a numeric vector}
    \item{\code{H35JRAZ03DF9NA}}{a numeric vector}
    \item{\code{H35JRAZ03DFE9W}}{a numeric vector}
    \item{\code{H35JRAZ03DFZXI}}{a numeric vector}
    \item{\code{H35JRAZ03DG0DK}}{a numeric vector}
    \item{\code{H35JRAZ03DGHF2}}{a numeric vector}
    \item{\code{H35JRAZ03DHRCR}}{a numeric vector}
    \item{\code{H35JRAZ03DIOTH}}{a numeric vector}
    \item{\code{H35JRAZ03DJAMF}}{a numeric vector}
    \item{\code{H35JRAZ03DJLV9}}{a numeric vector}
    \item{\code{H35JRAZ03DJU1V}}{a numeric vector}
    \item{\code{H35JRAZ03DKE4J}}{a numeric vector}
    \item{\code{H35JRAZ03DKRO9}}{a numeric vector}
    \item{\code{H35JRAZ03DLA3U}}{a numeric vector}
    \item{\code{H35JRAZ03DLJW2}}{a numeric vector}
    \item{\code{H35JRAZ03DLRGI}}{a numeric vector}
    \item{\code{H35JRAZ03DMDRM}}{a numeric vector}
    \item{\code{H35JRAZ03DMGA2}}{a numeric vector}
    \item{\code{H35JRAZ03DMGAO}}{a numeric vector}
    \item{\code{H35JRAZ03DO292}}{a numeric vector}
    \item{\code{H35JRAZ03DOWHJ}}{a numeric vector}
    \item{\code{H35JRAZ03DP0DM}}{a numeric vector}
    \item{\code{H35JRAZ03DQ55S}}{a numeric vector}
    \item{\code{H35JRAZ03DQK7F}}{a numeric vector}
    \item{\code{H35JRAZ03DR9O9}}{a numeric vector}
    \item{\code{H35JRAZ03DRJ6E}}{a numeric vector}
    \item{\code{H35JRAZ03DSA3E}}{a numeric vector}
    \item{\code{H35JRAZ03DSJQA}}{a numeric vector}
    \item{\code{H35JRAZ03DSR5L}}{a numeric vector}
  }
}
\source{
Elizabeth A. Green, Sarah W. Davies, Mikhail V. Matz, Monica Medina Next-generation sequencing reveals cryptic Symbiodinium diversity within Orbicella faveolata and Orbicella franksi at the Flower Garden Banks, Gulf of Mexico. PeerJ 2014 https://peerj.com/preprints/246/ 
}
\references{
Elizabeth A. Green, Sarah W. Davies, Mikhail V. Matz, Monica Medina Next-generation sequencing reveals cryptic Symbiodinium diversity within Orbicella faveolata and Orbicella franksi at the Flower Garden Banks, Gulf of Mexico. PeerJ 2014 https://peerj.com/preprints/246/ 
}
