\name{otuByAutocorr}
\alias{otuByAutocorr}
\title{
Selects OTUs for which MCMC-based parameter estimates are reliable.
}
\description{
Looks at autocorrelation among sampled parameter values.
}
\usage{
otuByAutocorr(model, data, ac.cut=0.1)
}
\arguments{
  \item{model}{
Model generated by mcmc.otu()
}
  \item{data}{
Dataset used to fit the model
}
  \item{ac.cut}{
Autocorrelation cutoff
}
}
\value{
Vector of OTU names for which autocorrelation is less than ac.cut argument.
}
\references{
Elizabeth A. Green, Sarah W. Davies, Mikhail V. Matz, Monica Medina Next-generation sequencing reveals cryptic Symbiodinium diversity within Orbicella faveolata and Orbicella franksi at the Flower Garden Banks, Gulf of Mexico. PeerJ 2014 https://peerj.com/preprints/246/ 
}
\author{
Mikhail V. Matz, University of Texas at Austin
<matz@utexas.edu>
}
