\name{panel.cor}
\alias{panel.cor}
\title{
accessory function for pairs() to display Pearson correlations
}
\description{
works as upper.panel or lower.panel argument of pairs() (package graphics).  
}
\usage{
panel.cor(x, y, digits=2, cex.cor) 
  }
\arguments{
  \item{x}{
x element of the pairs() matrix
}
  \item{y}{
y element of the pairs() matrix
}
  \item{digits}{
number of non-zero digits to leave at the end
}
  \item{cex.cor}{
scaling factor for displayed text
}
}
\references{
cannibalized from an example in ?pairs (package graphics)
}
